Finally fix the wayland flaky paste
Quote from WlDataOffer.receive:
> The receiving client reads from the read end of the pipe until EOF
  and then closes its end, at which point the transfer is complete.

The original idea was passing the stdin as the write pipe fd to the
receive(). However, sometimes the clipboard content was not pasted when
the richclip was called in neovim. That could be caused that, neovim
does not read the stdout fast enough. Since when the receive call
finishes, it ends the loop and quite the program.

To fix this, just follow the doc to create a dedicated pipe.
BTW, I was too lazy to do this at the beginning, since I was not aware of
nix-rust at that time -- I didn't want to get hands dirty with 'unsafe'.